Added salt to database

+++ unstash_secret.py
@@ -1,10 +1,17 @@
 from sql import execute_query
+from authenticate import authenticate_user
 
 def unstash_secret(id, force_delete):
 
-    if not force_delete:
-        user_input = input('Are you sure? (y/N)')
-        if user_input != 'y':
+        if (password := authenticate_user()):
+
+            if not force_delete:
+                user_input = input('Are you sure? (y/N)')
+                if user_input != 'y':
+                    return
+
+            execute_query(f'DELETE FROM secrets WHERE id={id}')
+            
+        else:
             return
 
-    execute_query(f'DELETE FROM secrets WHERE id={id}')
